New parameter: adapterType
A general idea is that the adapterType says about paralellism of modules-probe connection (how many modules are connected to a single element of the probe). The only exception is adapterType=0 which defines adapter type as the first adapter for the esaote probe.

adapterType = 0 -> 1st esaote adapter
adapterType = 2 -> ultrasonix adapter
adapterType = 1 -> 2nd esaote adapter (it is turned into -1 in the probeParams.m to pass the information that sth extra must be done)