Olympus 5L128 and vermon LA/20/180 probes added to dictionary.
According to data sheet HV for olympus probe is up to 180V for frequencies < 7.5MHz, and 115V for frequencies >10MHz. For safety reasons max voltage is now set to 115 V (for all frequencies).
HV for vermon probe is up to 30V (60Vpp according to data sheet). For safety reasons max voltage is now set to 15V.
Adapter for vermon probe (ATL/Philips) is yet not defined.