Reconstruction - new parameters
Warning: rx apodization in the reconstructRfImg.m & reconstructRFLin.m is determined by maxTang variable, which is calculated in these files. The new version uses user-defined values of rxTangLim (for the purpose of doppler processing).
The old files (ignoring the user defined rxTangLim) are being used when GPU processing is disabled.