Ref 205 doppler extension (#220)
This PR contains:
- CUDA kernels (CUDA MEX files) for: DDC, reconstruction, and Color Doppler processing;
- Color Doppler & Vector Doppler support.
New control script (Us4R_duplex.m) shows how to set new doppler-related parameters, and also some new b-mode related parameters. Old control scripts will work fine, but one has to remember, that there are some new parameters, which, if not set in the control script, will be set to default values. This is especially true for the bmodeRxTangLim parameter. The bmodeRxTangLim sets the RX tangent limits for the b-mode reconstruction, and its default value is [-0.5 0.5]. Its meaning is similar to maxTang parameter in old reconstruction functions (reconstructRfLin/reconstructRfImg). However, maxTang is calculated in those functions based on frequency, pitch, and speed of sound, while the bmodeRxTangLim is set by the user.


* CUDA MEX source file for LRI reconstruction

Initial version, no optimizations, PWI only.

* error message ID's

* Number of outputs can equal to 0.

Output then goes to "ans" variable

* CUDA MEX source file for Color Doppler

Initial version, no optimizations

* iqRaw2Lri - major changes

- more TX schemes included (STA, PWI)
- xElem moved to constant memory
- small performance improvements

* iqRaw2Lri - ownHypotf

* iqRaw2Lri - sinus & cosinus -> shared memory

* dopplerColor - major changes

- POWER added
- COLOR results are now in [rad] instead of [Hz], PRF is not required any more
- WALL CLUTTER FILTER added (simple rejection of constant component)

* iqRaw2Lri - major changes

- initial delay added
- rx tilting enabled
- textures used (but commented out, sth's wrong)
- rx apodization added (requires textures)
- sin & cos improved (requires textures)

* iqRaw2Lri - major changes

- texture used for sample picking & interpolation
- sin & cos calculated with __sincodf (no textures, no shared memory)
- rxApodization calculated as gaussian (no table look-up)
- some minor changes

* Some changes necessary for the ColorDoppler to work

* Minor fix in the boundary condition.

* DuplexDisplay - minor fixes

* Us4R - make rxTime a function of txPri

* iqRaw2Lri - checkData function for input validation

* iqRaw2Lri - give up the idea of multi-frame reconstruction

* iqRaw2Lri - small change in checkData function

* iqRaw2Lri - tiny change (style)

* iqRaw2Lri - bypass the limit of texture layers

* digitalDownConv - kernel for DDC

* uS4R - make use of the digitalDownConv kernel

There is a problem with a kernel, so its execution is commented out.

* Us4R - make use of the improved iqRaw2Lri kernel

No need to make the reconstruction in portions.

* Us4R - more universal step initialization

Before, it was valid only for certain filter types & orders.

* digitalDownConv - bug fix, kernel enabled in Us4R

* Make the final dynamic scale unaffected by the decimation

* iqRaw2Lri - rx tangent limits are constant scalars; txApod==0 case;

rx tangent limits are now scalars - they used to be vectors to enable duplex reconstruction (with varying rx tangents), but now the b-mode and color reconstructions are separated in Us4R.
txApod==0 now prevents unnecessary calculations and makes the output = nan (before, it was could be nan or zero)

* Us4R - major changes, doppler processing moved to dopplerColorImaging.m

- support for new rec parameters
- add path to CUDA kernels
- prepare reconstruction-related parameters for the use in GPU
- reorganize the reconstruction and color/vector-related code
- exclude the doppler processing to dopplerColorImaging.m

* Us4R - loading parameters during TX/RX for higher PRF.

* Reconstruction - new parameters

Warning: rx apodization in the reconstructRfImg.m & reconstructRFLin.m is determined by maxTang variable, which is calculated in these files. The new version uses user-defined values of rxTangLim (for the purpose of doppler processing).
The old files (ignoring the user defined rxTangLim) are being used when GPU processing is disabled.

* DuplexDisplay - major changes, final shape

* Us4R_doppler - set new parameters; new way of DuplexDisplay call

* Rename: Us4R_doppler.m -> Us4R_duplex.m

Co-authored-by: Piotr Jarosik <piotr.jarosik@us4us.eu>